+++ Changelog.md
@@ -1,10 +1,11 @@
 # Changelog
 
-## Release 13
+## Release 13 (next version)
 
 * Added `knGetAppSdk` for getting the app's target SDK and `knGetDeviceSdk` for getting the device's android version (sdk level)
 * Added `knJavaCommand` for calling smash hit's java commands from lua
 * Added `knGetDeviceHz` and `knSetFrameRate`
+* Better HTTP API; replaced `knHttpContentType()` with `knHttpGetHeader()`
 * Removed `knEnableReloading`; you can now just `knReload()` without enabling it
 * Fix a bug where the game would crash if using `knHttpRelease()` explicitly